requirements.txt
Flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
requests==2.31.0
Werkzeug==2.0.3
python-dotenv==1.0.0
boto3==1.26.137
botocore==1.29.137
pathlib==1.0.1
gevent==23.9.1 